feat(settings): move updates to its own settings tab
Extract auto-update toggle, feed URL, and check-now controls from
GeneralSettings into a dedicated UpdateSettings tab. Adds 'Updates'
entry to the SettingsModal TABS array.
